# backend/src/main/resources/openapi/customer-service.yaml
openapi: 3.0.3
info:
  title: Customer Service API
  version: 1.0.0
  description: API spec for customer-service backend
servers:
  - url: http://localhost:8082
paths:
...
  /api/v1/me:
    get:
      tags:
        - Auth
      operationId: getCurrentUser
      summary: Current user profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/admin/users:
    get:
      tags:
        - Admin
      operationId: listUsers
      summary: List all users (admin only)
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserAccountResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - Admin
      operationId: createUser
      summary: Create user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateUserRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAccountResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/admin/users/{id}:
    put:
      tags:
        - Admin
      operationId: updateUser
      summary: Update user role/status (admin only)
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAccountResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/admin/users/{id}/subscription:
    post:
      tags: [Admin]
      operationId: assignUserSubscription
      summary: Assign active package to user (admin)
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAssignSubscriptionRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/admin/packages:
    get:
      tags: [Admin]
      operationId: listAdminPackages
      summary: List service packages (admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PackageCatalogItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/admin/package-upgrade-requests:
    get:
      tags: [Admin]
      operationId: listPackageUpgradeRequests
      summary: List package upgrade requests (admin)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [PENDING, APPROVED, REJECTED]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PackageUpgradeRequestAdminDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/admin/package-upgrade-requests/{id}:
    patch:
      tags: [Admin]
      operationId: reviewPackageUpgradeRequest
      summary: Approve or reject upgrade request (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewPackageUpgradeRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageUpgradeRequestAdminDto"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
...
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    ErrorResponse:
      type: object
      required:
        - code
      properties:
        code:
          type: string
          example: INVALID_CREDENTIALS
        message:
          type: string
...
    UserProfile:
      type: object
      required:
        - id
        - userName
        - role
        - enabled
      properties:
        id:
          type: string
        userName:
          type: string
        role:
          type: string
          enum: [ADMIN, DEVELOP, USER]
        enabled:
          type: string
          enum: [ACTIVE, INACTIVE]
    UserAccountResponse:
      type: object
      required:
        - id
        - userName
        - role
        - enabled
      properties:
        id:
          type: string
        userName:
          type: string
        role:
          type: string
          enum: [ADMIN, DEVELOP, USER]
        enabled:
          type: string
          enum: [ACTIVE, INACTIVE]
    AdminCreateUserRequest:
      type: object
      required:
        - userName
        - password
        - role
        - enabled
      properties:
        userName:
          type: string
        password:
          type: string
          format: password
          minLength: 6
        role:
          type: string
          enum: [ADMIN, DEVELOP, USER]
        enabled:
          type: string
          enum: [ACTIVE, INACTIVE]
    AdminUpdateUserRequest:
      type: object
      properties:
        role:
          type: string
          enum: [ADMIN, DEVELOP, USER]
        enabled:
          type: string
          enum: [ACTIVE, INACTIVE]
        password:
          type: string
          format: password
          minLength: 6
...
    SubscriptionSummary:
      type: object
      required: [id, packageCode, tier, displayTitle, startDate, endDate, status]
      properties:
        id:
          type: string
          format: uuid
        packageCode:
          type: string
        tier:
          type: string
          enum: [BASIC, PRO]
        displayTitle:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          type: string
          enum: [ACTIVE, EXPIRED, CANCELLED]
...
    PackageCatalogItem:
      type: object
      required: [code, tier, label, quotaPosts, quotaImages, quotaVideos]
      properties:
        code:
          type: string
        tier:
          type: string
          enum: [BASIC, PRO]
        label:
          type: string
        quotaPosts:
          type: integer
        quotaImages:
          type: integer
        quotaVideos:
          type: integer
...
    AdminAssignSubscriptionRequest:
      type: object
      required: [packageCode]
      properties:
        packageCode:
          type: string
        displayTitle:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
    PackageUpgradeRequestAdminDto:
      type: object
      required: [id, userId, userName, fromPackageCode, toPackageCode, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        userName:
          type: string
        fromPackageCode:
          type: string
        toPackageCode:
          type: string
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        note:
          type: string
        adminNote:
          type: string
        createdAt:
          type: string
          format: date-time
        reviewedAt:
          type: string
          format: date-time
    ReviewPackageUpgradeRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [APPROVED, REJECTED]
        adminNote:
          type: string